Make ng2-uploader compatible with AoT Compiler
