+++ ng2-uploader.ts
@@ -1,20 +1,5 @@
-import { NgModule } from '@angular/core';
-import { NgFileDropDirective } from './src/directives/ng-file-drop';
-import { NgFileSelectDirective } from './src/directives/ng-file-select';
-import { Ng2Uploader } from './src/services/ng2-uploader';
+export * from './src/directives/ng-file-drop';
+export * from './src/directives/ng-file-select';
+export * from './src/services/ng2-uploader';
 
-@NgModule({
-   declarations: [
-     NgFileDropDirective,
-     NgFileSelectDirective
-   ],
-   providers: [
-     Ng2Uploader
-   ],
-   exports: [
-     NgFileDropDirective,
-     NgFileSelectDirective
-   ]
-})
-export class Ng2UploaderModule{}
-
+export { Ng2UploaderModule } from './src/module/ng2-uploader.module';
