+++ package.json
@@ -8,8 +8,8 @@
   "author": "Jan Kuri <jkuri88@gmail.com>",
   "scripts": {
     "clean": "./scripts/clean.sh",
-    "dev": "tsc --watch",
-    "prepublish": "tsc"
+    "dev": "ngc --watch",
+    "postinstall": "./node_modules/.bin/ngc -p tsconfig.json"
   },
   "repository": {
     "type": "git",
@@ -24,13 +24,20 @@
     "uploader"
   ],
   "devDependencies": {
-    "@angular/common": "^2.0.2",
-    "@angular/compiler": "^2.0.2",
-    "@angular/core": "^2.0.2",
-    "core-js": "^2.4.1",
+    "@angular/common": "^2.2.3",
+    "@angular/compiler": "^2.2.3",
+    "@angular/compiler-cli": "^2.2.3",
+    "@angular/core": "^2.2.3",
+    "@angular/platform-server": "^2.2.3",
+    "@angular/platform-browser": "^2.2.3",
+    "@angular/platform-browser-dynamic": "^2.2.3",
+    "@types/node": "6.0.51",
     "reflect-metadata": "^0.1.8",
-    "rxjs": "^5.0.0-beta.12",
-    "typescript": "^2.0.3",
+    "rxjs": "5.0.0-rc.4",
+    "typescript": "^2.0.10",
     "zone.js": "^0.6.25"
+  },
+    "dependencies": {
+    "@types/core-js": "^0.9.34"
   }
 }
