+++ scripts/clean.sh
@@ -3,3 +3,4 @@
 find . -name "*.js" -type f -not -path "./node_modules/*" -delete
 find . -name "*.js.map" -type f -not -path "./node_modules/*" -delete
 find . -name "*.d.ts" -type f -not -path "./node_modules/*" -delete
+find . -name "*.metadata.json" -type f -not -path "./node_modules/*" -delete
