+++ src/module/ng2-uploader.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { NgFileDropDirective } from './../directives/ng-file-drop';
+import { NgFileSelectDirective } from './../directives/ng-file-select';
+import { Ng2Uploader } from './../services/ng2-uploader';
+
+@NgModule({
+   declarations: [
+     NgFileDropDirective,
+     NgFileSelectDirective
+   ],
+   providers: [
+     Ng2Uploader
+   ],
+   exports: [
+     NgFileDropDirective,
+     NgFileSelectDirective
+   ]
+})
+export class Ng2UploaderModule{}
+
